Refactor: move out utility funcs, remove from module level export
