+++ esp_data/backends/webdataset_backend.py
@@ -1,198 +1,16 @@
 """WebDataset implementation of the StreamingBackend protocol."""
 
-import io
-import json
 from typing import Any, Callable, Iterator
 
-import numpy as np
-import soundfile as sf
 import webdataset as wds
 
-from esp_data.io import AnyPathT, PureCloudPath, anypath, filesystem_from_path
+from esp_data.io import AnyPathT, anypath
 
 from .protocol import StreamingBackend
+from .webdataset_utils import audio_decoder
 
 
-def audio_decoder(data: dict, dtype: str = "float32", format: str = "FLAC") -> dict[str, Any]:
-    """Decode audio data from a WebDataset sample.
-
-    Parameters
-    ----------
-    data: dict
-        The sample containing audio data in WebDataset format
-    dtype: str
-        The data type of the decoded audio data (default: "float32")
-    format: str
-        The format of the audio data (default: "FLAC")
-
-    Returns
-    -------
-    dict
-        Dictionary containing the decoded audio data and metadata.
-
-    Raises
-    ------
-    ValueError
-        If the sample does not contain an audio key ending with .flac, .wav, etc.
-    """
-    audio_key = next((k for k in data if k.endswith(f".{format.lower()}")), None)
-    if not audio_key:
-        raise ValueError("Sample must contain an audio key ending with .flac, .wav, etc.")
-
-    audio_buffer = io.BytesIO(data[audio_key])
-    audio_data, samplerate = sf.read(audio_buffer, dtype=dtype)
-
-    # Reconstruct sample
-    sample = {}
-    sample["audio"] = audio_data
-    sample["sample_rate"] = samplerate
-    md = json.loads(data.get("metadata.json", "{}").decode("utf-8"))
-    sample.update(md)
-
-    return sample
-
-
-def audio_encoder(
-    sample: dict[str, Any],
-    sample_rate: int = 16000,
-    dtype: str = "float32",
-    format: str = "FLAC",
-) -> dict[str, Any]:
-    """Encode audio data in the sample to a specific format.
-
-    Parameters
-    ----------
-    sample: dict[str, Any]
-        The sample containing audio data
-    sample_rate: int
-        The sample rate of the audio data
-    dtype: str
-        The data type of the audio data (default: "float32")
-    format: str
-        The format to encode the audio data to (e.g., "WAV", "FLAC", "OGG")
-        Default is "FLAC".
-
-    Returns
-    -------
-    dict
-        Dictionary containing the encoded audio data and metadata
-        in the WebDataset format.
-
-    Raises
-    ------
-    ValueError
-        If the sample does not contain an "audio" key with audio data.
-    """
-    if "audio" not in sample:
-        raise ValueError("Sample must contain 'audio' key with audio data")
-
-    data_out = {}
-    audio_buffer = io.BytesIO()
-    # Convert audio data to the specified format
-    if isinstance(sample["audio"], (list, tuple)):
-        # If audio is a list or tuple, convert to numpy array
-        sample["audio"] = np.array(sample["audio"], dtype=dtype)
-    elif isinstance(sample["audio"], np.ndarray):
-        # If audio is already a numpy array, ensure it's the correct dtype
-        sample["audio"] = sample["audio"].astype(dtype)
-
-    sf.write(audio_buffer, sample["audio"], sample_rate, format=format)
-
-    data_out[f"audio.{format.lower()}"] = audio_buffer.getvalue()
-
-    # Add metadata (without audio)
-    sample = {k: v for k, v in sample.items() if k != "audio"}  # Remove audio key from metadata
-    data_out["metadata.json"] = json.dumps(sample, indent=2).encode("utf-8")
-    return data_out
-
-
-def json_encoder(
-    sample: dict[str, Any],
-    indent: int = 2,
-) -> dict[str, Any]:
-    """Encode a sample to JSON format.
-
-    Parameters
-    ----------
-    sample: dict[str, Any]
-        The sample to encode
-    indent: int
-        Indentation level for JSON (default: 2)
-
-    Returns
-    -------
-    dict
-        Dictionary containing the encoded sample in JSON format.
-    """
-    json_data = json.dumps(sample, indent=indent).encode("utf-8")
-    return {"sample.json": json_data}
-
-
-def json_decoder(
-    data: dict[str, Any],
-) -> dict[str, Any]:
-    """Decode a sample from JSON format.
-
-    Parameters
-    ----------
-    data: dict[str, Any]
-        The sample containing JSON data
-
-    Returns
-    -------
-    dict
-        Dictionary containing the decoded sample.
-
-    Raises
-    ------
-    ValueError
-        If the sample does not contain a "sample.json" key.
-    """
-    if "sample.json" not in data:
-        raise ValueError("Sample must contain 'sample.json' key with JSON data")
-
-    json_data = json.loads(data["sample.json"].decode("utf-8"))
-    return json_data
-
-
-def make_file_opener_for_wds(
-    file_path: str | AnyPathT,
-    mode: str = "wb",
-    block_size: int = 1024 * 1024 * 100,
-) -> Callable:
-    """Make a file opener function for WebDataset.
-
-    If local path, create parent dirs if needed.
-
-    Arguments
-    ---------
-    file_path: str | AnyPathT
-        The file path to open
-    mode: str
-        The mode in which to open the file (default: "wb")
-    block_size: int
-        Block size for WebDataset (default: 100 MB)
-
-    Returns
-    -------
-    Callable
-        A function that opens the file in the specified mode
-        or a file object if the path is local.
-    """
-    path_obj = anypath(file_path)
-
-    if not isinstance(path_obj, PureCloudPath):
-        # Local filesystem - create parent dirs if needed
-        parent_dir = path_obj.parent
-        parent_dir.mkdir(parents=True, exist_ok=True)
-        return open(str(path_obj), mode=mode)
-    else:
-        # Remote filesystem (GCS, R2, etc.)
-        fs = filesystem_from_path(str(path_obj))
-        return fs.open(str(path_obj.no_prefix), mode=mode, block_size=block_size)
-
-
-def load_webdataset(
+def _load_webdataset(
     path: str | AnyPathT,
     file_pattern: str = "shard*tar",
     data_processor: Callable | None = audio_decoder,
@@ -352,7 +170,7 @@ class WebDatasetBackend(StreamingBackend):
         WebDatasetBackend
             Wrapped WebDataset backend
         """
-        dataset = load_webdataset(
+        dataset = _load_webdataset(
             path,
             file_pattern=file_pattern,
             data_processor=data_processor,
